Small Screen Layout Changes
Buttons and login bar are now centre when viewed on a small screen
(below 414px width).
Removed the display of the login bar for now until I work out how to
have it the way it was without using inline-block.
Added a gitignore thing so that github will hopefully ignore the HTML
index page I'm using to test everything